t/25-sql.t: Fix the DBD::DBM problems
Thanks to Birmingham.pm for investigation and a proposed patch.

Further investigation showed that SQL::Statement 1.20 didn't trigger
the problem, but all versions further did (the latest at the time of
this commit is 1.26).

The fix for DBD::DBM is to force using DBI::SQL::Nano when the
available version of SQL::Statement is known to cause the bug.