Workaround problem with NextRelease plugin
The NextRelease plugin requires a a PAUSE username. The username can be
provided by a ~/.dzil/config.ini file, but if the file
is missing it will ask for a PAUSE id. This will not work in an
noninteractive automated GitHub action workflow.

It is not currently clear to me how to fix this issue, so for now I just
omit using the NextRelease plugin.